+++ test/jason/support.clj
@@ -1,0 +1,6 @@
+(ns jason.support
+  (:require
+    [clojure.string :as string]))
+
+(defn multiline-str [& args]
+  (string/join "\n" args))